recent

+++ src/kaosendrequest/alimtalk.go
@@ -5,55 +5,72 @@ import (
 	"database/sql"
 	"encoding/json"
 	"fmt"
-	kakao "mycs/src/kakaojson"
-	config "mycs/src/kaoconfig"
-	databasepool "mycs/src/kaodatabasepool"
-
-	//"io/ioutil"
-	//	"net"
-	//"net/http"
+	"context"
 	"strconv"
 	s "strings"
 	"sync"
 	"time"
+
+	kakao "mycs/src/kakaojson"
+	config "mycs/src/kaoconfig"
+	databasepool "mycs/src/kaodatabasepool"
 )
 
 var atprocCnt int
 
-func AlimtalkProc() {
+func AlimtalkProc(user_id string, ctx context.Context) {
 	atprocCnt = 1
-	 
+	config.Stdlog.Println(user_id, " - 알림톡 프로세스 시작 됨 ")
+
 	for {
-		if atprocCnt <=5 {
-			var startNow = time.Now()
-			var group_no = fmt.Sprintf("%02d%02d%02d%09d", startNow.Hour(), startNow.Minute(), startNow.Second(), startNow.Nanosecond())
-	
-			updateRows, err := databasepool.DB.Exec("update DHN_REQUEST_AT set send_group = '" + group_no + "' where send_group is null limit " + strconv.Itoa(config.Conf.SENDLIMIT))
-	
-			if err != nil {
-				config.Stdlog.Println("알림톡 send_group Update 오류 : ", err)
-			}
-	
-			rowcnt, _ := updateRows.RowsAffected()
-	
-			if rowcnt > 0 {
-				config.Stdlog.Println("알림톡 발송 처리 시작 ( ", group_no, " ) : ", rowcnt, " 건 ")
-				atprocCnt++
-				go atsendProcess(group_no)
-	
+		if atprocCnt <=10 {
+			
+			select {
+				case <- ctx.Done():
+			    config.Stdlog.Println(user_id, " - 알림톡 process가 10초 후에 종료 됨.")
+			    time.Sleep(10 * time.Second)
+			    config.Stdlog.Println(user_id, " - 알림톡 process 종료 완료")
+			    return
+			default:
+				var count sql.NullInt64
+				cnterr := databasepool.DB.QueryRowContext(ctx, "SELECT count(1) AS cnt FROM DHN_REQUEST_AT WHERE send_group IS NULL AND userid=?", user_id).Scan(&count)
+				
+				if cnterr != nil && cnterr != sql.ErrNoRows {
+					config.Stdlog.Println(user_id, "- 알림톡 DHN_REQUEST Table - select 오류 : " + cnterr.Error())
+					time.Sleep(10 * time.Second)
+				} else {
+					if count.Valid && count.Int64 > 0 {		
+						var startNow = time.Now()
+						var group_no = fmt.Sprintf("%02d%02d%02d%09d", startNow.Hour(), startNow.Minute(), startNow.Second(), startNow.Nanosecond())
+						
+						updateRows, err := databasepool.DB.ExecContext(ctx, "update DHN_REQUEST_AT set send_group = ? where send_group is null and userid = ?  limit ?", group_no, user_id, strconv.Itoa(config.Conf.SENDLIMIT))
+				
+						if err != nil {
+							config.Stdlog.Println(user_id," - 알림톡 send_group Update 오류 : ", err)
+						}
+				
+						rowcnt, _ := updateRows.RowsAffected()
+				
+						if rowcnt > 0 {
+							config.Stdlog.Println(user_id, " - 알림톡 발송 처리 시작 ( ", group_no, " ) : ", rowcnt, " 건 ")
+							atprocCnt++
+							go atsendProcess(group_no, user_id)
+				
+						}
+					}
+				}
 			}
 		}
 	}
-
 }
 
-func atsendProcess(group_no string) {
+func atsendProcess(group_no, user_id string) {
 	var db = databasepool.DB
 	var conf = config.Conf
 	var stdlog = config.Stdlog
 	var errlog = config.Stdlog
 
-	reqsql := "select * from DHN_REQUEST_AT where send_group = '" + group_no + "'  "
+	reqsql := "select * from DHN_REQUEST_AT where send_group = '" + group_no + "' and userid = '" + user_id +"'"
 
 	reqrows, err := db.Query(reqsql)
 	if err != nil {
@@ -382,11 +399,12 @@ title) values %s`
 		resinsValues = nil
 	}
 
-	db.Exec("delete from DHN_REQUEST_AT where send_group = '" + group_no + "'")
+	db.Exec("delete from DHN_REQUEST_AT where send_group = '" + group_no + "' and userid = '" + user_id +"'")
 
+	atprocCnt--
+	
 	stdlog.Println("알림톡 발송 처리 완료 ( ", group_no, " ) : ", procCount, " 건 ( Proc Cnt :", atprocCnt, ")")
 	
-	atprocCnt--
 }
 
 func sendKakaoAlimtalk(reswg *sync.WaitGroup, c chan<- resultStr, alimtalk kakao.Alimtalk, temp resultStr) {
